Use .PHONY for targets (solve problems with objects/ dir)
The targets in this Makefile are being used "generically" as labels
(think: names of things that can run a sequence of commands) and not as
actual file-based dependencies as Make prefers.  This is OK to do, but
in the case of the objects/ directory, leads to a real-world problem.
Example scenario:

  git clone ...
  cd bake
  make
  {...fails due to cc not being installed on the user's system,...}
  {...however mkdir -p objects/ has been run successfully...}
  {...user installs compiler...}
  make
  {...does nothing, because objects/ directory has been created...}

This happens because the target "objects" happens to also be the same
name as the directory where *.o files are stored.  Once the directory
is created, that dependency (re: "all: objects") is fulfilled, thus make
does nothing.

The .PHONY directive is used to tell Make what is not a real file-based
dependency and instead just a generic target name:
https://www.gnu.org/software/make/manual/html_node/Phony-Targets.html

In general though, a proper Makefile with correct file dependencies
(both objs, cpp, and hpp files) should really be written.